requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via discord-py
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
attrs==24.3.0
    # via aiohttp
azure-cognitiveservices-speech==1.42.0
    # via -r requirements.in
cachetools==5.5.0
    # via google-auth
certifi==2024.12.14
    # via requests
cffi==1.17.1
    # via pynacl
charset-normalizer==3.4.1
    # via requests
discord-ext-voice-recv==0.4.2a145
    # via -r requirements.in
discord-py==2.4.0
    # via
    #   -r requirements.in
    #   discord-ext-voice-recv
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
google-auth==2.37.0
    # via google-genai
google-genai==0.5.0
    # via -r requirements.in
idna==3.10
    # via
    #   requests
    #   yarl
mss==10.0.0
    # via -r requirements.in
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
numpy==2.2.2
    # via
    #   opencv-python
    #   scipy
opencv-python==4.11.0.86
    # via -r requirements.in
pillow==11.1.0
    # via
    #   -r requirements.in
    #   google-genai
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pyaudio==0.2.14
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.10.5
    # via google-genai
pydantic-core==2.27.2
    # via pydantic
pynacl==1.5.0
    # via discord-py
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via google-genai
rsa==4.9
    # via google-auth
scipy==1.15.1
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
urllib3==2.3.0
    # via requests
websockets==14.2
    # via google-genai
yarl==1.18.3
    # via aiohttp
